Windowing System HW 1
Windowing Systems Programming assignment:

Outline:
In this assignment, students will create a GUI that displays several
labels and a button, and then
connect a simple callback to the button.

Interface
Students should create a GUI containing the following elements:
• a label in each corner of the GUI (one in the lower-left, one in the
lower-right, one in the
upper-left, and one in the upper-right)
• a push-button in the center of the GUI
Each component should maintain its positioning as the window is resized
– that is, the pushbutton
should stay in the center of the screen, the four labels should stay in
their respective corners,
and none of them should be “stretched out” as the window resizes.

Events:
Students should attach an event to the push-button so that, when it is
clicked, it prints (to the
console) the number of times that the button has been clicked. (Students
will probably have to add
an additional field to their GUI class, so that it can keep track of the
number of times the button has
been clicked.)