Windowing System HW 2
Windowing Systems Programming assignment:

Outline:
In this assignment, students will create a GUI that displays several
radio buttons, a check button
and a normal push-button.

Interface:
Students should create a GUI containing the following elements:
• three radio buttons in the top-left of the window
• three radio buttons in the top-right of the window, in a different
group
• a check-box in the bottom-left corner of the window
• a push-button in the bottom-right corner of the window
Each component should maintain its positioning as the window is resized,
and none of them
should be stretched out as the window resizes.

Events:
Students should arrange it so that the check-box controls whether or not
the push-button is
enabled. (When the check-box is checked, the push-button should be
enabled; when the check-box
is unchecked, the push-button should be disabled.)
Students should arrange it so that, when the push-button is clicked, the
application prints out
which radio-buttons are selected. (Bear in mind that there are two
groups, so two radio buttons
might be selected. Also, depending on how your GUI is set up, it might
be possible for no radio
buttons to be selected.)
Students should include a comment in each callback, stating what event
on what object the
callback is attached to, and what class the callback is a member-method
of.