Windowing System HW 4
Windowing Systems Programming assignment:

Interface:
Students should create a GUI containing the following elements:
• a combo-box and a button in the top-let corner
• a combo-box and a button in the top-right corner
• a tab manager in the bottom-left corner
• a stack-panel in the bottom-right corner
There should not be any wasted space in the interface; the tab manager
and stack-panel should
expand to take up any space that is added to the layout as the GUI is
resized. (The tab manager and
stack-panel do not have to be equally sized; one could have a fixed
width while the other resized
freely, for example.)
The combo-box in the top-left corner will be used to select between the
tab-manager and stack
panel, while the combo-box in the top-right corner will be used to
select between a message-box
and save-file dialog. Each should have appropriate values.

Events:
Students should arrange it so that, when the button next to the top-left
combo box (the “combobox
or stack panel” box) is clicked, one of two things happens, depending on
the selected value in
the combo-box:
• if the tab manager is selected, then a new page is added to the tab
manager. For the content
of the new page, the student should use a text box wrapped in a
scroll-viewer.
• If the stack-panel is selected, then a new item is added to the stack
panel. For a new item,
the student should use a label
Students should arrange it so that, when the button next to the
top-right combo box (the “message
or dialog” box) is clicked, one of two things happens, depending on the
selected value in the
combo-box:
• if “message box” is selected, then a message dialog-box should be
displayed. The message
box should have at least two options, say “cancel” and “OK”, and the
program should print
which is selected.
• If “file dialog” is selected, then a file dialog-box should be
displayed. If the user selects a
file, the file-path should be printed.